+++ Console/IPDatabase/COPYRIGHT.txt
@@ -1,0 +1,1 @@
+Database and Contents Copyright (c) 2019 MaxMind, Inc.
