+++ Console/packages.config
@@ -3,7 +3,11 @@
   <package id="Microsoft.CodeAnalysis.FxCopAnalyzers" version="2.9.3" targetFramework="net452" />
   <package id="Microsoft.CodeAnalysis.VersionCheckAnalyzer" version="2.9.3" targetFramework="net452" />
   <package id="Microsoft.CodeQuality.Analyzers" version="2.9.3" targetFramework="net452" />
+  <package id="MaxMind.Db" version="2.4.0" targetFramework="net452" />
+  <package id="MaxMind.GeoIP2" version="3.0.0" targetFramework="net452" />
+  <package id="Microsoft.CSharp" version="4.4.1" targetFramework="net452" />
   <package id="Microsoft.Maps.MapControl.WPF" version="1.0.0.3" targetFramework="net45" userInstalled="true" />
   <package id="Microsoft.NetCore.Analyzers" version="2.9.3" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net452" />
   <package id="Microsoft.NetFramework.Analyzers" version="2.9.3" targetFramework="net452" />
 </packages>
